+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/AutoBrewer.java
@@ -1,6 +1,5 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.electric.machines;
 
-import city.norain.slimefun4.SlimefunExtended;
 import com.xzavier0722.mc.plugin.slimefun4.storage.util.StorageCacheUtils;
 import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
 import io.github.thebusybiscuit.slimefun4.api.items.SlimefunItemStack;
@@ -21,7 +20,6 @@ import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.PotionMeta;
-import org.bukkit.potion.PotionData;
 import org.bukkit.potion.PotionType;
 
 /**
@@ -144,11 +142,7 @@ public class AutoBrewer extends AContainer implements NotHopperable {
 
     @ParametersAreNonnullByDefault
     private @Nullable ItemStack brew(Material input, Material potionType, PotionMeta potion) {
-        if (SlimefunExtended.isAtLeast(1, 20, 2)) {
-            return brewPostBasePotionType(input, potionType, potion);
-        } else {
-            return brewPreBasePotionType(input, potionType, potion);
-        }
+        return brewPostBasePotionType(input, potionType, potion);
     }
 
     private PotionType getExtendedPotionType(PotionType type) {
@@ -170,6 +164,10 @@ public class AutoBrewer extends AContainer implements NotHopperable {
     @ParametersAreNonnullByDefault
     private ItemStack brewPostBasePotionType(Material input, Material potionType, PotionMeta potion) {
         PotionType type = potion.getBasePotionType();
+        if (type == null) {
+            return null;
+        }
+
         if (type == PotionType.WATER) {
             if (input == Material.FERMENTED_SPIDER_EYE) {
                 potion.setBasePotionType(PotionType.WEAKNESS);
@@ -189,14 +187,20 @@ public class AutoBrewer extends AContainer implements NotHopperable {
                 potion.setBasePotionType(fermented);
                 return new ItemStack(potionType);
             }
-        } else if (input == Material.REDSTONE && type.isExtendable() && !type.isUpgradeable()) {
+        } else if (input == Material.REDSTONE && type.isExtendable()) {
             // Fixes #3390 - Potions can only be either extended or upgraded. Not both.
-            potion.setBasePotionType(getExtendedPotionType(type));
-            return new ItemStack(potionType);
-        } else if (input == Material.GLOWSTONE_DUST && type.isUpgradeable() && !type.isExtendable()) {
+            PotionType extended = getExtendedPotionType(type);
+            if (extended != null) {
+                potion.setBasePotionType(extended);
+                return new ItemStack(potionType);
+            }
+        } else if (input == Material.GLOWSTONE_DUST && type.isUpgradeable()) {
             // Fixes #3390 - Potions can only be either extended or upgraded. Not both.
-            potion.setBasePotionType(getStrengthenPotionType(type));
-            return new ItemStack(potionType);
+            PotionType strengthened = getStrengthenPotionType(type);
+            if (strengthened != null) {
+                potion.setBasePotionType(strengthened);
+                return new ItemStack(potionType);
+            }
         } else if (input == Material.GUNPOWDER && potionType == Material.POTION) {
             return new ItemStack(Material.SPLASH_POTION);
         } else if (input == Material.DRAGON_BREATH && potionType == Material.SPLASH_POTION) {
@@ -210,49 +214,6 @@ public class AutoBrewer extends AContainer implements NotHopperable {
             }
         }
 
-        return null;
-    }
-
-    @ParametersAreNonnullByDefault
-    @SuppressWarnings("deprecration")
-    private ItemStack brewPreBasePotionType(Material input, Material potionType, PotionMeta potion) {
-        PotionData data = potion.getBasePotionData();
-        PotionType type = data.getType();
-        if (type == PotionType.WATER) {
-            if (input == Material.FERMENTED_SPIDER_EYE) {
-                potion.setBasePotionData(new PotionData(PotionType.WEAKNESS, false, false));
-                return new ItemStack(potionType);
-            } else if (input == Material.NETHER_WART) {
-                potion.setBasePotionData(new PotionData(PotionType.AWKWARD, false, false));
-                return new ItemStack(potionType);
-            } else if (potionType == Material.POTION && input == Material.GUNPOWDER) {
-                return new ItemStack(Material.SPLASH_POTION);
-            } else if (potionType == Material.SPLASH_POTION && input == Material.DRAGON_BREATH) {
-                return new ItemStack(Material.LINGERING_POTION);
-            }
-        } else if (input == Material.FERMENTED_SPIDER_EYE) {
-            PotionType fermented = fermentations.get(type);
-
-            if (fermented != null) {
-                potion.setBasePotionData(new PotionData(fermented, data.isExtended(), data.isUpgraded()));
-                return new ItemStack(potionType);
-            }
-        } else if (input == Material.REDSTONE && type.isExtendable() && !data.isUpgraded()) {
-            // Fixes #3390 - Potions can only be either extended or upgraded. Not both.
-            potion.setBasePotionData(new PotionData(type, true, false));
-            return new ItemStack(potionType);
-        } else if (input == Material.GLOWSTONE_DUST && type.isUpgradeable() && !data.isExtended()) {
-            // Fixes #3390 - Potions can only be either extended or upgraded. Not both.
-            potion.setBasePotionData(new PotionData(type, false, true));
-            return new ItemStack(potionType);
-        } else if (type == PotionType.AWKWARD) {
-            PotionType potionRecipe = potionRecipes.get(input);
-
-            if (potionRecipe != null) {
-                potion.setBasePotionData(new PotionData(potionRecipe, false, false));
-                return new ItemStack(potionType);
-            }
-        }
         return null;
     }