+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/magical/staves/StormStaff.java
@@ -89,7 +89,7 @@ public class StormStaff extends LimitedUseItem {
         }
 
         if (p.getGameMode() != GameMode.CREATIVE) {
-            FoodLevelChangeEvent event = new FoodLevelChangeEvent(p, p.getFoodLevel() - 4);
+            FoodLevelChangeEvent event = new FoodLevelChangeEvent(p, p.getFoodLevel() - 4, item);
             Bukkit.getPluginManager().callEvent(event);
 
             if (!event.isCancelled()) {